fix some bugs

+++ app/src/main/java/com/blogspot/mido_mymall/ui/my_cart/MyCartUtil.kt
@@ -9,6 +9,6 @@ interface MyCartUtil {
     fun deleteItem(position:Int)
 
 
-    fun getTotalAmount(totalAmount:Int)
+    fun getTotalAmount(totalAmount:Double)
 
 }
